Refactor main Nyabot class
